Give the sampling validation its own mixin
The client grew past its length budget with the multi round-trip branch's
sampling checks. The handling of a server's sampling request and the shape
rules it applies move unchanged into Client::SamplingValidation, which the
client includes; every method stays private.
